tests/test_gRPC_service/service_embedding/requirements.txt
fastmcp
grpcio
grpcio-tools
protobuf
llama-index
langchain
langchain-community
langchain-openai
sentence-transformers
transformers
torch
numpy
pydantic
python-dotenv
asyncio
uvloop
